+++ secrets_env/console/__init__.py
@@ -2,5 +2,6 @@ import secrets_env.console.commands.completion
 import secrets_env.console.commands.config
 import secrets_env.console.commands.keyring
 import secrets_env.console.commands.run
+import secrets_env.console.commands.shell
 
 from .core import entrypoint